Fix telemetry: accept float uptimeSec from Pi runtime
Pi sends uptime as a float (e.g. 123.45) but Zod schema required int.
Relaxed to z.number() to accept both.

+++ web/lib/validators.ts
@@ -97,7 +97,7 @@ export const telemetrySchema = z.object({
     ramTotalMb: z.number().int().min(0).nullable().optional(),
     tempC: z.number().nullable().optional(),
     ip: z.string().trim().max(64).nullable().optional(),
-    uptimeSec: z.number().int().min(0).nullable().optional(),
+    uptimeSec: z.number().min(0).nullable().optional(),
     internet: z.boolean().default(false),
   }),
 })
